all functions for reading config from flash moved to CustomDevice.cpp
